add product feed infra
