+++ packages/infrastructure/src/mcp/saleor/graphql/fragments/ProductFeedFragment.graphql
