+++ packages/infrastructure/src/mcp/saleor/graphql/queries/ProductsFeedQuery.graphql
